requirements.txt
streamlit
pandas
plotly.express
plotly.graph_objects
plotly.subplots
datetime
numpy
scikit-learn